backend/requirements.txt
alembic==1.14.0
altgraph==0.17.4
asgiref==3.8.1
bakong-khqr==0.2.8
bcrypt==4.2.1
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.9
Django==5.1.5
et_xmlfile==2.0.0
filelock==3.16.1
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-MySQLdb==2.0.0
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
Mako==1.3.8
MarkupSafe==3.0.2
mutagen==1.47.0
mysql-connector==2.2.9
mysql-connector-python==9.1.0
mysqlclient==2.2.6
openpyxl==3.1.5
packaging==24.1
pefile==2023.2.7
pillow==11.0.0
platformdirs==4.3.6
psycopg2==2.9.10
pycryptodomex==3.21.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyMySQL==1.1.1
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
PySide6==6.8.0.1
PySide6_Addons==6.8.0.1
PySide6_Essentials==6.8.0.1
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytube==15.0.0
pywin32-ctypes==0.2.3
requests==2.32.3
shiboken6==6.8.0.1
simple-websocket==1.1.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
termcolor==2.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.28.1
websockets==13.1
Werkzeug==3.1.3
wsproto==1.2.0
youtube-dl==2021.12.17
yt-dlp==2024.10.7
